Add a session expiry modal
This checks periodically for sessions at or near expiry using JS, then
shows a modal telling the user to log in in another tab (with a
target=_blank link). This should help reduce frustration when filling
out a long form (e.g. newsletter) and hitting submit, only to find the
Meetup OAuth token has expired.

Would be nice to have token refresh but I don't have time to implement
it.